trty fix really bad and inconsitent work of ai.. chatgpt40.. it started of with varations.  its predicted the geometric movement of formeladyed based on kinematrics and shpereical symmetry.  moving towardss reversible vertel fix point dyanamics and adding tons to the thread on manyh topics, it lost the plot.   the coding is terrible. now body coordinates,use of magic numbers, no pattern, no structure, repeat numbers, hacked html code.  ?
use of 3 is ok but performance is weak. links to three js are live.. it not cmpiled or packaged.  the LLM is not ready from primetime as its inconsistent and the tooling is horrible.

the expectations are too high and frustrations result from its moments of genius and completetent to beign stupid.

wll be movigj nto  offline LLMs if possible or wait it out as it has toom  much konwledges to fit on a pc.

its get but my money is on device makers and not NVIDIA its really hot and its all Qualcom , nokia and possiblly nueromophic..  Intel needs a new CEO.. spriked neoronal chips are part of how our brained wrosk. it uses al 3... d0x86.  riskx, 2d , reversbiel, exact bitwise, adn   high temp 3d, llow temp2d, spiked neural without best tomography, and symbolic processing. its a chimera brain thats how real thigns are.   but the experience of asking slow online chatgpt4 to fix it. fix it fixt.. no .. dont make me do anything.. i raht just do it all than copilot at this point.     or have ai agent chains do it all.. or have it start off a prject then i fix it.  its promising but   way too frustrating.. ill use it less.   if it were on device and fastre and more consider with zero safety crap.   mabye its good. now is questionalble value adn prive/ vaule propostion is earth boiling.. need to pull alll stops on this crap.    people can (dshouid be able to)  spot fake ais. if they cant they need to go outside more.  people are the best liars, but they shouldntn need to be and usually are detected, truth is power